refactor: iriondaengdaeng 디렉토리 내 컴포넌트 파일들을 components 디렉토리로 이동
